feat: add reviews refactor: make ids Long type refactor: remove inheritance refactor: add logging to controllers

+++ src/main/java/ru/yandex/practicum/filmorate/service/impl/DirectorServiceImpl.java
@@ -36,18 +36,18 @@ public class DirectorServiceImpl implements DirectorService {
     }
 
     @Override
-    public Director get(Integer id) {
+    public Director get(Long id) {
         checkDirectorExists(id);
         return storage.get(id);
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         checkDirectorExists(id);
         storage.delete(id);
     }
 
-    private void checkDirectorExists(int id) {
+    private void checkDirectorExists(Long id) {
         if (!storage.containsInStorage(id)) {
             throw new NotFoundException(String.format("Director with id = %d not found", id));
         }
